Refactor retriever builder tests
