+++ x-pack/plugin/rank-rrf/src/main/java/org/elasticsearch/xpack/rank/rrf/RRFRetrieverBuilder.java
@@ -245,7 +245,7 @@ public final class RRFRetrieverBuilder extends CompoundRetrieverBuilder<RRFRetri
     }
 
     @Override
-    protected RetrieverBuilder doRewrite(QueryRewriteContext ctx) {
+    public RetrieverBuilder doRewrite(QueryRewriteContext ctx) {
         RetrieverBuilder rewritten = this;
 
         ResolvedIndices resolvedIndices = ctx.getResolvedIndices();
